+++ ServerDevcommands/Commands/Bind.cs
@@ -1,9 +1,7 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection.Emit;
 using HarmonyLib;
-using UnityEngine;
 namespace ServerDevcommands;
 ///<summary>Adds support for mouse wheel binding and other features.</summary>
 [HarmonyPatch]